% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cik_search.R
\name{cik_search}
\alias{cik_search}
\title{SEC CIK Search}
\usage{
cik_search(company)
}
\arguments{
\item{company}{Search term to search for CIK}
}
\value{
A dataframe with one row per company with 
       Includes the following columns -
 \itemize{
   \item cik
   \item company_href
   \item company_name
 }
}
\description{
Provides access to the SEC CIK search tool from
\href{https://www.sec.gov/edgar/searchedgar/cik.htm}{here}
}
\examples{
cik_search("cloudera")
}
