% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_details.R
\name{company_details}
\alias{company_details}
\title{SEC Company Details}
\usage{
company_details(x, ownership = FALSE, type = "", before = "",
  count = 40, page = 1)
}
\arguments{
\item{x}{URL to a SEC filing index page}
}
\value{
A dataframe with all the parsed meta-info on the filing
}
\description{
The SEC generates a html page as an index for every filing it receives containing
all the metainformation about the filing. We extract 3 main types of information: 
\itemize{
 \item Company Information - Filing date, accepted date, etc.
 \item Filings - Companies included in the filing
}
}
\details{
For a company, there is typically a single filer and no funds, but many filings
for funds get more complicated - e.g. 400+ funds with 100's of companies

NOTE: This can get process intensive for large fund pages. If you don't need all
components, try just usning filing_info
}
