% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{company_href}
\alias{company_href}
\title{Company URL for a CIK}
\usage{
company_href(cik, ownership = FALSE, atom = FALSE)
}
\arguments{
\item{cik}{Company code}

\item{ownership}{(default: FALSE) boolean for inclusion of company change
filings}

\item{atom}{(default: FALSE) if the link should be to the atom XML feed}
}
\value{
A string with URL requested
}
\description{
Given a CIK, provide a link to the company information page.
}
\examples{
company_href("0000037912")
}
