% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_info.R
\name{company_info}
\alias{company_info}
\title{SEC Company Info}
\usage{
company_info(ticker)
}
\arguments{
\item{ticker}{Either a stock symbol (for the 10,000 largest companies) or CIK
code}
}
\value{
a dataframe with all SEC company information
}
\description{
Fetches basic information on a given company from the SEC site
}
