% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_information.R
\name{company_information}
\alias{company_information}
\title{SEC Company Info}
\usage{
company_information(x)
}
\arguments{
\item{x}{Either a stock symbol (for the 10,000 largest companies) or CIK
code}
}
\value{
a dataframe with all SEC company information
}
\description{
Fetches basic information on a given company from the SEC site
}
