% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{effectiveness}
\alias{effectiveness}
\title{SEC Notice of Effectiveness}
\usage{
effectiveness()
}
\value{
a data.frame with each row as a submission with the following
  columns:
  \describe{
    \item{registration_number}{}
    \item{file_href}{}
    \item{registrant}{}
    \item{registrant_href}{}
    \item{filing_date}{}
    \item{effective_date}{}
    \item{division}{}
    \item{type}{}
 }
}
\description{
Returns the current Noice of Effectiveness from the most recently completed
business day from
\href{https://www.sec.gov/cgi-bin/browse-edgar?action=geteffect}{here}
}
\details{
You can also see the same filings going further back by using
`latest_filings()` specifying the type = "EFFECT"
}
\examples{
effectiveness()
}
