% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing_documents.R
\name{filing_documents}
\alias{filing_documents}
\title{SEC Filing Documents}
\usage{
filing_documents(href)
}
\arguments{
\item{href}{URL to a SEC filing index page}
}
\value{
A dataframe with all the documents in the filing along with their
   meta info
}
\description{
SEC Filing Documents
}
