% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing_information.R
\name{filing_information.character}
\alias{filing_information.character}
\title{rdname filing_information}
\usage{
\method{filing_information}{character}(x)
}
\description{
rdname filing_information
}
