% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing_information.R
\name{filing_information}
\alias{filing_information}
\title{SEC Filing Information}
\usage{
filing_information(href)
}
\arguments{
\item{href}{URL to a SEC filing index page}
}
\value{
A dataframe with all the parsed meta-info on the filing
}
\description{
The SEC generates a html page as an index for every filing it receives containing
all the metainformation about the filing.
}
\details{
Information returned:
 * type
 * description
 * accession_number
 * filing_date
 * accepted_date
 * documents
 * period_date
 * changed_date
 * effective_date
 * filing_bytes
Not all details are valid for all filings, but the column will always be present

If you know you're going to want all the details of a filing, including documents
funds and filers, look at `filing_details`
}
