% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_xml.R
\name{map_xml}
\alias{map_xml}
\title{Map XML Entries}
\usage{
map_xml(
  doc,
  entries_xpath,
  parts,
  trim = c(),
  integers = c(),
  date_format = ""
)
}
\arguments{
\item{doc}{An XML document}

\item{entries_xpath}{an xpath locator for all starting points}

\item{parts}{a list in the form column name = xpath locator}

\item{trim}{a list of columns that need to have whitespace trimmed}

\item{integers}{a list of columns that need to converted to integers}

\item{date_format}{how dates should be parsed}
}
\value{
A dataframe with one row per entry and columns from parts
}
\description{
Extracts a dataframe from an xml document.
}
\keyword{internal}
