% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_submission.R
\name{parse_submission}
\alias{parse_submission}
\title{Parse Submission}
\usage{
parse_submission(uri)
}
\arguments{
\item{uri}{- URL to a SEC complete submiision text file}
}
\value{
a tibble with one row per document, the content nested in TEXT
}
\description{
SEC filings are sent in a SGML file - this parses that master submission
into component documents, with content lines in list column 'TEXT'.
}
\details{
*NOTE:* non-text documents are uuencoded and likely are missing needed
padding on lines to allow uudecode to work.
}
\examples{
parse_submission(paste0('https://www.sec.gov/Archives/edgar/data/',
                 '37996/000003799617000084/0000037996-17-000084.txt'))
}
