% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_codes.R
\docType{data}
\name{sic_codes}
\alias{sic_codes}
\title{SIC Codes}
\format{A data frame with 1005 rows and 6 variables:
\describe{
  \item{sic}{Standard Industrial Classification}
  \item{industry}{Name of industry}
  \item{division_id}{Letter code for the division}
  \item{division}{Name of the division}
  \item{major}{Name of the major group, identified by 1st 2 digits of the sic}
  \item{group}{Name of the group, identified by the 1st 3 digits of the sic}
}}
\source{
\url{https://www.osha.gov/pls/imis/sic_manual.html}

\url{https://www.sec.gov/info/edgar/siccodes.htm}
}
\usage{
sic_codes
}
\description{
SIC code table with structure.
}
\keyword{datasets}
