% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing_details.R
\name{filing_details}
\alias{filing_details}
\title{SEC Filing Details}
\usage{
filing_details(href)
}
\arguments{
\item{href}{URL to a SEC filing index page}
}
\value{
A dataframe with all the parsed meta-info on the filing
}
\description{
The SEC generates a html page as an index for every filing it receives containing
all the metainformation about the filing. We extract 3 main types of information: 
\itemize{
 \item Filing Information - Filing date, accepted date, etc.
 \item Documents - All the documents included in the filing
 \item Filers - Companies included in the filing
 \item Funds - Funds included in the filing
}
}
\details{
For a company, there is typically a single filer and no funds, but many filings
for funds get more complicated - e.g. 400+ funds with 100's of companies

NOTE: This can get process intensive for large fund pages. If you don't need all
components, try just usning filing_info
}
