% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing_filers.R
\name{filing_filers}
\alias{filing_filers}
\title{SEC Filing Included Filers}
\usage{
filing_filers(x)
}
\arguments{
\item{x}{URL to a SEC filing index page}
}
\value{
A dataframe with all the filers in the filing along with their info
}
\description{
SEC Filing Included Filers
}
\examples{
filing_details("https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/0000712515-17-000063-index.htm")
}
