% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing_funds.R
\name{filing_funds}
\alias{filing_funds}
\title{SEC Filing Funds}
\usage{
filing_funds(x)
}
\arguments{
\item{x}{URL to a SEC filing index page}
}
\value{
A dataframe with all the funds associated with a given filing
}
\description{
SEC Filing Funds
}
