% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing_funds.R
\name{filing_funds}
\alias{filing_funds}
\alias{filing_funds.character}
\alias{filing_funds.xml_node}
\title{SEC Filing Funds}
\usage{
filing_funds(x)

\method{filing_funds}{character}(x)

\method{filing_funds}{xml_node}(x)
}
\arguments{
\item{x}{URL to a SEC filing index page}
}
\value{
A dataframe with all the funds associated with a given filing
}
\description{
SEC Filing Funds
}
\examples{
\donttest{
# Typically you'd get the URL from one of the search functions
x <- paste0("https://www.sec.gov/Archives/edgar/data/",
            "933691/000119312517247698/0001193125-17-247698-index.htm")
try(filing_funds(x))
}
}
